+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.2] - 2024-12-07
+## [1.2.1] - 2024-12-11
+
+### Added
+
+- Modal for context tag repr items, to see what items belong under that tag
+
+### Fixed
+
+- Update for StardewUI HEAD (changed GridItemLayout.Length ctor)
+- Byproducts heading not appearing on modal
+- Weedses
+
+## [1.2.0] - 2024-12-07
 
 ### Added